Brazilian Farmers

From Valente (1995) “In the mid-1960s, Rogers and others conducted an ambitious ‘three country study’ to determine influences on adoption of farm practices in Nigeria, India and Brazil. [...] Only in Brazil, and only for hybrid corn, did adoption of the innovation reach more than a small proportion of the farmers.”

brfarmers

Format

A data frame with 692 rows and 148 columns:

village

village number

idold

respondent id

age

respondent's age

liveout

Lived outside of community

visits

# of visits to large city

contact

# of contacts with relatives

coop

membership in coop

orgs

membership in organizations

patry

Patriarchalism score

liter

Literate

news1

# of newspapers or mags pr mon

subs

subscribe to news

radio1

Own radio

radio2

Frequency radio listening

radio3

program preference

tv

frequency Tv viewing

movie

freq movie attendance

letter

freq letter writing

source

total # of sources used for ag

practA

Ever used practice A

practB

Ever used practice B

practC

Ever used practice C

practD

Ever used practice D

practE

Ever used practice E

practF

Ever used practice F

practG

Ever used practice G

practH

Ever used practice H

practI

Ever used practice I

practJ

Ever used practice J

practK

Ever used practice K

practL

Ever used practice L

yrA

A year of adoption

yrB

B year of adoption

yrC

C year of adoption

yrD

D year of adoption

yrE

E year of adoption

yrF

F year of adoption

yrG

G year of adoption

yrH

H year of adoption

yrI

I year of adoption

yrJ

J year of adoption

yrK

K year of adoption

yrL

L year of adoption

curA

A Current use

curB

B Current use

curC

C Current use

curD

D Current use

curE

E Current use

curF

F Current use

curG

G Current use

curH

H Current use

curI

I Current use

curJ

J Current use

curK

K Current use

curL

L Current use

srce1

Source of aware in A

timeA

Years ago 1st aware

src2

Source of more info on A

src3

Most influential source

use

use during trial stage

total

total # of practices adopted

futatt

Future attitude

achiev

Achievement Score

attcred

Attitude toward credit

littest

Score on functional literacy t

acarcomm

Communication with ACAR repres

econk

Economic knowledge

caact

recognize any change agent act

hfequip

# of home & farm equips owned

politk

political knowledge score

income

income

land1

total land area in pasture

land2

total land area planted

cows

# of cows giving milk

land3

total land owned

respf

respondent named as friend

respa

respondent named as ag adv

resppa

respondent named for practic A

resppb

respondent named for practic B

resppc

respondent named for practic C

poly

polymorphic OL for 3 practices

respl

respondent named for loan

resppi

resp named for price info

repsccp

resp named for coop comm proj

counter

counterfactuality score

opinion

opinionness score

school

years of schooling by resp

pk1

political know 1

pk2

political know 2

pk3

political know 3

pk4

political know 4

pk5

political know 5

innovtim

innovativeness time

adoptpct

adoption percent

discon

# of practices discontinued

mmcred

Mass media credibility

trust

Trust

stusincn

Status inconsistency

nach

N achievement motivation

attcred2

Attitude toward credit

risk

Risk taking

socpart

Social participate

patriarc

patriarchy

crdit2

attit to credit for product

visicit

visitin cities

nondep

non-dependence on farming

oltotal

OL total 7 items t-score

innov

overall innovativeness score

icosmo

cosmo index

immexp

mass media exposure index

iempath

empathy index

iach5

achievement motivation index 5

iach7

achievement motivation index 7

ipk

political knowledge index

immc

mass media credibililty index

iol

OL index

yr

Actual Year of Adoption

fs

--- MISSING INFO ---

ado

Time of Adoption

tri

Triangular values used as appro

hlperc

high low percent of diffusion

hlperc1

--- MISSING INFO ---

new

new or old villages

card1

card number

sour1

Source: radio

sour2

Source: TV

sour3

Source: Newpaper

sour4

Source: Magazine

sour5

Source: ACAR Bulletin

sour6

Source: Agronomist

sour7

Source: Neighbor

sourc6

--- MISSING INFO ---

adopt

--- MISSING INFO ---

net31

nomination friend 1

net32

nomination friend 2

net33

nomination friend 3

net21

nomination influential 1

net22

nomination influential 2

net23

nomination influential 3

net11

nomination practice A

net12

nomination practice B

net13

nomination practice C

net41

nomination coop comm proj

id

--- MISSING INFO ---

commun

Number of community

toa

Time of Adoption

test

--- MISSING INFO ---

study

Number of study in Valente (1995)

Source

The Brazilian Farmers data were collected as part of a USAID-funded study of farming practicing in the three countries, India, Nigeria, and Brazil. There was only one wave of data that contained survey questions regarding social networks, and only in Brazil did diffusion of the studied farming innovations reach an appreciable saturation level- that was for hybrid seed corn. The data were stored along with hundreds of other datasets by the University of Wisconsin library and I, Tom Valente, paid a fee to have the disks mailed to me in the early 1990s.

Details

The dataset has 692 respondents (farmers) from 11 communities. Collected during 1966, it spans 20 years of farming pracitices.

References

Rogers, E. M., Ascroft, J. R., & Röling, N. (1970). Diffusion of Innovation in Brazil, Nigeria, and India. Unpublished Report. Michigan State University, East Lansing.

Valente, T. W. (1995). Network models of the diffusion of innovations (2nd ed.). Cresskill N.J.: Hampton Press.

See also