Korean Family Planning

From Valente (1995) “Scholars at Seoul National University's School of Public Health (Park, Chung, Han & Lee, 1974) collected data on the adoption of family planning methods among all married women of child-bearing age 25 in Korea villages in 1973 (N = 1,047).”

Format

A data frame with 1,047 rows and 432 columns:

village

Village of residence

id

Respondent ID number

recno1

Card number NA

studno1

Study number NA

area1

Village of residence

id1

Respondent ID number

nmage1

Number males age 0

nmage2

Number males age 0-4

nmage3

Number males age 5-9

nmage4

Number males age 10-14

nmage5

Number males age 15-19

nmage6

Number males age 20-24

nmage7

Number males age 25-29

nmage8

Number males age 30-34

nmage9

Number males age 35-39

nmage10

Number males age 40-44

nmage11

Number males age 45-49

nmage12

Number males age 50-54

nmage13

Number males age 55-59

nmage14

Number males age 60-64

nmage15

Number males age 65-69

nmage16

Number males age 70-74

nmage17

Number males age 75-79

nmage18

Number males age 80+

nfage1

Number females age 0

nfage2

Number females age 0-4

nfage3

Number females age 5-9

nfage4

Number females age 10-14

nfage5

Number females age 15-19

nfage6

Number females age 20-24

nfage7

Number females age 25-29

nfage8

Number females age 30-34

nfage9

Number females age 35-39

nfage10

Number females age 40-44

nfage11

Number females age 45-49

nfage12

Number females age 50-54

nfage13

Number females age 55-59

nfage14

Number females age 60-64

nfage15

Number females age 65-69

nfage16

Number females age 70-74

nfage17

Number females age 75-79

nfage18

Number females age 80+

pregs

total pregnancies

pregs1

number normal deliveries

pregs2

number of induced abortions

pregs3

number of spontaneous abortions

pregs4

number of still births

pregs5

number of deaths after live birth

pregs6

currently pregnant

sons

number of sons

daughts

number of daughters

planning

Ever heard of FP or birth control

loop1

Awareness of Loop

loop2

Detailed knowledge of Loop

loop3

Attitudes toward Loop

loop4

Knowledge of Loop used by neighbors

loop5

Knowledge of place of service for Loop

pill1

Awareness of Pill

pill2

Detailed knowledge of Pill

pill3

Attitudes toward Pill

pill4

Knowledge of Pill used by neighbors

pill5

Knowledge of place of service for Pill

vase1

Awareness of Vasectomy

vase2

Detailed knowledge of Vasectomy

vase3

Attitudes toward Vasectomy

vase4

Knowledge of Vasectomy used by neighbors

vase5

Knowledge of place of service for Vasectomy

cond1

Awareness of Condoms

cond2

Detailed knowledge Condoms

cond3

Attitudes toward Condoms

cond4

Knowledge of Condoms used by neighbors

cond5

Knowledge of place of service for Condoms

rhyt1

Awareness of Rhythm

rhyt2

Detailed knowledge Rhythm

rhyt3

Attitudes toward Rhythm

rhyt4

Knowledge of Rhythm used by neighbors

bbt1

Awareness of Basic Body Temperature

bbt2

Detailed knowledge Basic Body Temperature

bbt3

Attitudes toward BBT

recno2

Record Number NA

studno2

Study Number NA

area2

village number

id2

id number

bbt4

Knowledge of BBT used by neighbors

diap1

Awareness of Diaphragm

diap2

Detailed knowledge Diaphragm

diap3

Attitudes toward Diaphragm

diap4

Knowledge of Diaphragm used by neighbors

with1

Awareness of Withdrawal

with2

Detailed knowledge Withdrawal

with3

Attitudes toward Withdrawal

with4

Knowledge of Withdrawal used by neighbors

tuba1

Awareness of Tubal Ligation

tuba2

Detailed knowledge TL

tuba3

Attitudes toward TL

tuba4

Knowledge of TL used by neighbors

fp1

Experience with an FP practice

fp2

Reasons for not practicing

fp3

What would you do if problem was solved

fp4

Any other reason for not practicing

fp5

Reasons for practicing

fp6

time between decision and adoption

fp7

reasons for time lag

fp8

Ever discontinued practicing

fp9

Reasons for discontinuing

fp10

Attitude toward FP

child1

Ideal number of sons

child2

Ideal number of daughters

child3

Ideal number of children regardless of sex

child4

what do if kept having girls

comop1

Spousal communication on # of children

comop2

Spousal communication on FP

comop3

Consensus on opinion between couple

comop4

What was the difference

comop5

Opinion on who should practice

comop6

Different opinions on who should practice

comop7

Who should make final decision

comop8

Residence in old age

net11

Neighbors talk to about FP- 1

net12

Neighbors talk to about FP- 2

net13

Neighbors talk to about FP- 3

net14

Neighbors talk to about FP- 4

net15

Neighbors talk to about FP- 5

famawe1

Family members of FP Practice

famawe2

Parents awareness of FP Practice

famawe3

How did parents-in-law become aware

famawe4

How did parents become aware

famawe5

How did husband become aware

advic1

Advice given to neighbors where to go

advic2

Advice given on method

advic3

Ever met persons who give advice on FP

advic4

Credibility of person advising on FP

advic5

Counter advice given to others

rumor1

Rumors on Loop

rumor2

Rumors on Pill

rumor3

Rumors on Vasectomy

rumor4

Rumors on Condom

rumor5

Rumors on Tuballigation

media1

Possession of Radio

media2

Possession of TV

media3

Subscription to Newspaper

media4

Subscription to Happy Home

media5

Subscription to other magazine

media6

Radio exposure to FP

media7

TV exposure to FP

media8

Daily paper exposure to FP

media9

Happy Home exposure to FP

media10

Magazine exposure to FP

media11

Movie or slide exposure to FP

media12

Poster exposure to FP

media13

Pamphlet exposure to FP

media14

FP Meeting exposure to FP

recno3

Record number NA

studno3

Study number NA

area3

village

id3

id

media15

Public lecture exposure to FP

media16

Mobile van exposure to FP

media17

Neighbors exposure to FP

media18

Workers home visiting exposure to FP

media19

Husband exposure to FP

club1

Awareness of clubs in community

club2

Membership in club

club3

Reasons for not becoming a member

club4

Feeling of necessity of club

club5

Visit of mobile van to area

club6

Service received from van

club7

Decision-making on FP on # children

club8

Decision-making on important goods

club9

Decision-making on childrens discipline

club10

Decision making on purchase wife clothes

net21

Closest neighbor most frequently met

n1adv

Advice received from neighbor 1

n1prac

practice of FP by neighbor 1

net22

Closest neighbor person 2

n2adv

Advice received from neighbor 2

n2prac

Practice of FP by neighbor 2

net23

Closest neighbor person 3

n3adv

Advice received from neighbor 3

n3prac

Practice of FP by neighbor 3

net24

Closest neighbor 4

n4adv

Advice received from neighbor 4

n4prac

Practice of FP by neighbor 4

net25

Closest neighbor 5

n5adv

Advice received from neighbor 5

n5prac

Practice of FP by neighbor 5

stand

Standard living of above neighbors

educ

Education level of named neighbors

net31

Advice on FP sought from 1

net32

Advice on FP sought from 2

net33

Advice on FP sought from 3

net34

Advice on FP sought from 4

net35

Advice on FP sought from 5

net41

Information provided on FP by 1

net42

Information provided on FP by 1

net43

Information provided on FP by 1

net44

Information provided on FP by 1

net45

Information provided on FP by 1

net51

Seek advice on induced abortion 1

net52

Seek advice on induced abortion 2

net53

Seek advice on induced abortion 3

net54

Seek advice on induced abortion 4

net55

Seek advice on induced abortion 5

age

Age of respondent

agemar

Age at first marriage

recno4

Rec no NA

studno4

Study no NA

area4

village

id4

id

net61

Advice on health sought from 1

net62

Advice on health sought from 2

net63

Advice on health sought from 3

net64

Advice on health sought from 4

net65

Advice on health sought from 5

net71

Advice on purchase of goods 1

net72

Advice on purchase of goods 2

net73

Advice on purchase of goods 3

net74

Advice on purchase of goods 4

net75

Advice on purchase of goods 5

net81

Advice on childrens education 1

net82

Advice on childrens education 2

net83

Advice on childrens education 3

net84

Advice on childrens education 4

net85

Advice on childrens education 5

rfampl1

Advice on FP sought by 1

rfampl2

Advice on FP sought by 2

rfampl3

Advice on FP sought by 3

rfampl4

Advice on FP sought by 4

rfampl5

Advice on FP sought by 5

rfampll

Leadership score - indegree FP

rabort1

Advice on abortion sought by 1

rabort2

Advice on abortion sought by 2

rabort3

Advice on abortion sought by 3

rabort4

Advice on abortion sought by 4

rabort5

Advice on abortion sought by 5

rabortl

Leadership score - indegree abortion

rhealth1

Advice on health sought by 1

rhealth2

Advice on health sought by

rhealth3

Advice on health sought by

rhealth4

Advice on health sought by

rhealth5

Advice on health sought by

rhealthl

Leadership score - indegree health

recno5

rec no NA

studno5

study no NA

area5

village

id5

id

rgoods1

Advice on purchases sought by 1

rgoods2

Advice on purchases sought by 2

rgoods3

Advice on purchases sought by 3

rgoods4

Advice on purchases sought by 4

rgoods5

Advice on purchases sought by 5

rgoodsl

Leadership score - indegree purchases

reduc1

Advice on education sought by 1

reduc2

Advice on education sought by 2

reduc3

Advice on education sought by 3

reduc4

Advice on education sought by 4

reduc5

Advice on education sought by 5

reducl

Leadership score - indegree education

hub1

Husbands friend 1

hub2

Husbands friend 2

hub3

Husbands friend 3

hub4

Husbands friend 4

hub5

Husbands friend 5

hubed

Husbands education

wifeed

Wifes education

wiferel

Wifes religion

hubocc

Husbands occupation

wifeocc

Wifes occupation

know1

Can you insert a loop yourself

know2

Can you remove it alone

know3

Can a man use a loop

know4

How long can a loop be used

know5

Which doctor

know6

Doctor or nurse

know7

Oral pill method

know8

Can men take pills

know9

Long term use

know10

Time required for vasectomy

know11

Does vasectomy = castration

know12

Can any doctor do vasectomies

pref1

Who prefer use: Husband or wife

pref2

Reasons for preferring FP practice by wife

pref3

Reasons for preferring FP practice by husband

ageend

Ideal age to end childbearing

cfp

Current status of FP

cfatt1

Husbands attitude

cfatt2

In-laws attitude

cfatt3

Own parents attitude

cbyr

Start of period from year

cbmnth

Start of period from month

ceyr

End of period year

cemnth

End of period month

clngth

Length of period

cawe1

FP contact

cawe2

Awareness of contraceptive method at the time

cawe3

Awareness of service site

cawe4

Credibiilty

recno6

rec no NA

studno6

study no NA

area6

village

id6

id

fpt1

FP Status time 1

fatt1t1

Husbands attitude T1

fatt2t1

In-laws attitude T1

fatt3t1

Own parents attitude T1

byrt1

Start of Time 1 from year

lngtht1

Length of Time 1

awe1t1

FP Contact Time 1

awe2t1

Methods known at Time 1

awe3t1

Knowledge of service sites Time 1

awe4t1

Credibility of service site Time 1

fpt2

FP Status time 2

fatt1t2

Husbands attitude T2

fatt2t2

In-laws attitude T2

fatt3t2

Own parents attitude T2

byrt2

Start of Time 2 from year

lngtht2

Length of Time 2

awe1t2

FP Contact Time 2

awe2t2

Methods known at Time 2

awe3t2

Knowledge of service sites Time 2

awe4t2

Credibility of service site Time 2

fpt3

FP Status time 3

fatt1t3

Husbands attitude T3

fatt2t3

In-laws attitude T3

fatt3t3

Own parents attitude T3

byrt3

Start of Time 3 from year

lngtht3

Length of Time 3

awe1t3

FP Contact Time 3

awe2t3

Methods known at Time 3

awe3t3

Knowledge of service sites Time 3

awe4t3

Credibility of service site Time 3

fpt4

FP Status time 4

fatt1t4

Husbands attitude T4

fatt2t4

In-laws attitude T4

fatt3t4

Own parents attitude T4

byrt4

Start of Time 4 from year

lngtht4

Length of Time 4

awe1t4

FP Contact Time 4

awe2t4

Methods known at Time 4

awe3t4

Knowledge of service sites Time 4

awe4t4

Credibility of service site Time 4

fpt5

FP Status time 5

fatt1t5

Husbands attitude T5

fatt2t5

In-laws attitude T5

fatt3t5

Own parents attitude T5

byrt5

Start of Time 5 from year

lngtht5

Length of Time 5

awe1t5

FP Contact Time 5

awe2t5

Methods known at Time 5

awe3t5

Knowledge of service sites Time 5

awe4t5

Credibility of service site Time 5

fpt6

FP Status time 6

fatt1t6

Husbands attitude T6

fatt2t6

In-laws attitude T6

fatt3t6

Own parents attitude T6

byrt6

Start of Time 6 from year

lngtht6

Length of Time 6

awe1t6

FP Contact Time 6

awe2t6

Methods known at Time 6

awe3t6

Knowledge of service sites Time 6

awe4t6

Credibility of service site Time 6

recno7

rec no NA

studno7

study no NA

area7

village

id7

id

fpt7

FP Status time 7

fatt1t7

Husbands attitude T7

fatt2t7

In-laws attitude T7

fatt3t7

Own parents attitude T7

byrt7

Start of Time 7 from year

lngtht7

Length of Time 7

awe1t7

FP Contact Time 7

awe2t7

Methods known at Time 7

awe3t7

Knowledge of service sites Time 7

awe4t7

Credibility of service site Time 7

fpt8

FP Status time 8

fatt1t8

Husbands attitude T8

fatt2t8

In-laws attitude T8

fatt3t8

Own parents attitude T8

byrt8

Start of Time 8 from year

lngtht8

Length of Time 8

awe1t8

FP Contact Time 8

awe2t8

Methods known at Time 8

awe3t8

Knowledge of service sites Time 8

awe4t8

Credibility of service site Time 8

fpt9

FP Status time 9

fatt1t9

Husbands attitude T9

fatt2t9

In-laws attitude T9

fatt3t9

Own parents attitude T9

byrt9

Start of Time 9 from year

lngtht9

Length of Time 9

awe1t9

FP Contact Time 9

awe2t9

Methods known at Time 9

awe3t9

Knowledge of service sites Time 9

awe4t9

Credibility of service site Time 9

fpt10

FP Status time 10

fatt1t10

Husbands attitude T10

fatt2t10

In-laws attitude T10

fatt3t10

Own parents attitude T10

byrt10

Start of Time 10 from year

lngtht10

Length of Time 10

awe1t10

FP Contact Time 10

awe2t10

Methods known at Time 10

awe3t10

Knowledge of service sites Time 10

awe4t10

Credibility of service site Time 10

fpt11

FP Status time 11

fatt1t11

Husbands attitude T11

fatt2t11

In-laws attitude T11

fatt3t11

Own parents attitude T11

byrt11

Start of Time 11 from year

lngtht11

Length of Time 11

awe1t11

FP Contact Time 11

awe2t11

Methods known at Time 11

awe3t11

Knowledge of service sites Time 11

awe4t11

Credibility of service site Time 11

fpt12

FP Status time 12

fatt1t12

Husbands attitude T12

fatt2t12

In-laws attitude T12

fatt3t12

Own parents attitude T12

byrt12

Start of Time 12 from year

lngtht12

Length of Time 12

awe1t12

FP Contact Time 12

awe2t12

Methods known at Time 12

awe3t12

Knowledge of service sites Time 12

awe4t12

Credibility of service site Time 12

ado

adopt times years converted to 1=63

ado1

ado2

ado3

commun

Village number

toa

Time of Adoption

study

Study (for when multiple diff studies used)

Source

The Korean Family Planning data were stored on a Vax tape that Rogers had given to Marc Granovetter who then gave it to his colleague Roland Soong (see Granovetter & Soong, 1983). Granovetter instructed Song to send the tape to me and I had it loaded on the Vax machine at USC in 1990 and was able to download the data to a PC. The first two datasets were acquired for my dissertation (Valente, 1991) and the third added as I completed my book on Network Models of the Diffusion of Innovations (Valente, 1995; also see Valente, 2005).

Details

The dataset has 1,047 respondents (women) from 25 communities. Collected during 1973 it spans 11 years of data.

References

Everett M. Rogers, & Kincaid, D. L. (1981). Communication Networks: Toward a New Paradigm for Research. (C. Macmillan, Ed.). New York; London: Free Press.

Valente, T. W. (1995). Network models of the diffusion of innovations (2nd ed.). Cresskill N.J.: Hampton Press.

See also