Medical Innovation

From Valente (1995) “Coleman, Katz and Menzel from Columbia University's Bureau of Applied Research studied the adoption of tetracycline by physiciams in four Illinois communities in 1954.[...] Tetracycline was a powerful and useful antibiotic just introduced in the mid-1950s”

Format

A data frame with 125 rows and 59 columns:

city

city id

id

sequential respondent id

detail

detail man

meet

meetings, lectures, hospitals

coll

colleagues

attend

attend professional meets

proage

professional age

length

lenght of reside in community

here

only practice here

science

science versus patients

position

position in home base

journ2

journal subscriptions

paadico

Percent alter adoption date imp

ado

adoption month 1 to 18

thresh

threshold

ctl

corrected tl tl-exp level

catbak

category 1-init 2-marg 3-low tl

sourinfo

source of information

origid

original respondent id

adopt

adoption date 1= 11/53

recon

reconstructed med innov

date

date became aware

info

information source

most

most important info source

journ

journals

drug

drug houses

net1_1

advisor nomination1

net1_2

advisor nomination2

net1_3

advisor nomination3

net2_1

discuss nomination1

net2_2

discuss nomination2

net2_3

discuss nomination3

net3_1

friends nomination1

net3_2

friends nomination2

net3_3

friends nomination3

nojourn

number of pro journals receive

free

free time companions

social

med discussions during social

club

club membership

friends

friends are doctors

young

young patients

nonpoor

nonpoverty patients

office

office visits

house

house calls

tend

tendency to prescribe drugs

reltend

relative tendency to prescribe

perc

perceived drug competition

proximty

physical proximity to other doc

home

home base hospital affiliation

special

specialty

belief

belief in science

proage2

profesional age 2

presc

prescription prone

detail2

contact with detail man

dichot

dichotomous personal preference

expect

adoption month expected

recall

recalls adopting

commun

Number of community

toa

Time of Adoption

study

Number of study in Valente (1995)

Source

The Medical Innovation data were stored in file cabinets in a basement building at Columbia University. Ron Burt (1987) acquired an NSF grant to develop network diffusion models and retrieve the original surveys and enter them into a database. He distributed copies of the data on diskette and sent one to me, Tom Valente, and I imported onto a PC environment.

Details

The collected dataset has 125 respondents (doctors), and spans 17 months of data collected in 1955. Time of adoption of non-adopters has been set to month 18 (see the manual entry titled Difussion Network Datasets).

References

Coleman, J., Katz, E., & Menzel, H. (1966). Medical innovation: A diffusion study (2nd ed.). New York: Bobbs-Merrill

Valente, T. W. (1995). Network models of the diffusion of innovations (2nd ed.). Cresskill N.J.: Hampton Press.

See also