Summary of diffnet objects

Summary of diffnet objects

# S3 method for diffnet
summary(
  object,
  slices = NULL,
  no.print = FALSE,
  skip.moran = FALSE,
  valued = getOption("diffnet.valued", FALSE),
  ...
)

Arguments

object
An object of class diffnet.

slices
Either an integer or character vector. While integer vectors are used as indexes, character vectors are used jointly with the time period labels.

no.print
Logical scalar. When TRUE suppress screen messages.

skip.moran
Logical scalar. When TRUE Moran's I is not reported (see details).

valued
Logical scalar. When TRUE weights will be considered. Otherwise non-zero values will be replaced by ones.

...
Further arguments to be passed to approx_geodesic.

Value

A data frame with the following columns:

adopt

Integer. Number of adopters at each time point.

cum_adopt

Integer. Number of cumulative adopters at each time point.

cum_adopt_pcent

Numeric. Proportion of comulative adopters at each time point.

hazard

Numeric. Hazard rate at each time point.

density

Numeric. Density of the network at each time point.

moran_obs

Numeric. Observed Moran's I.

moran_exp

Numeric. Expected Moran's I.

moran_sd

Numeric. Standard error of Moran's I under the null.

moran_pval

Numeric. P-value for the observed Moran's I.

Details

Moran's I is calculated over the cumulative adoption matrix using as weighting matrix the inverse of the geodesic distance matrix. All this via moran. For each time period t, this is calculated as:

 m = moran(C[,t], G^(-1))

Where C[,t] is the t-th column of the cumulative adoption matrix, G^(-1) is the element-wise inverse of the geodesic matrix at time t, and moran is netdiffuseR's moran's I routine. When skip.moran=TRUE Moran's I is not reported. This can be useful for both: reducing computing time and saving memory as geodesic distance matrix can become large. Since version 1.18.0, geodesic matrices are approximated using approx_geodesic which, as a difference from geodist from the sna package, and distances from the igraph package returns a matrix of class dgCMatrix (more details in approx_geodesic).

See also

Examples

data(medInnovationsDiffNet)
summary(medInnovationsDiffNet)

#> Diffusion network summary statistics
#> Name     : Medical Innovation
#> Behavior : Adoption of Tetracycline
#> -----------------------------------------------------------------------------
#>  Period   Adopters   Cum Adopt. (%)   Hazard Rate   Density   Moran's I (sd)  
#> -------- ---------- ---------------- ------------- --------- ---------------- 
#>        1         11        11 (0.09)             -      0.02  0.07 (0.03) **  
#>        2          9        20 (0.16)          0.08      0.02  0.04 (0.03)     
#>        3          9        29 (0.23)          0.09      0.02 -0.03 (0.03)     
#>        4         11        40 (0.32)          0.11      0.02 -0.02 (0.03)     
#>        5         11        51 (0.41)          0.13      0.02 -0.06 (0.03)     
#>        6         11        62 (0.50)          0.15      0.02 -0.02 (0.03)     
#>        7         13        75 (0.60)          0.21      0.02 -0.00 (0.03)     
#>        8          7        82 (0.66)          0.14      0.02  0.01 (0.03)     
#>        9          4        86 (0.69)          0.09      0.02  0.01 (0.03)     
#>       10          1        87 (0.70)          0.03      0.02  0.01 (0.03)     
#>       11          5        92 (0.74)          0.13      0.02  0.02 (0.03)     
#>       12          3        95 (0.76)          0.09      0.02  0.02 (0.03)     
#>       13          3        98 (0.78)          0.10      0.02  0.01 (0.03)     
#>       14          4       102 (0.82)          0.15      0.02  0.04 (0.03)     
#>       15          4       106 (0.85)          0.17      0.02  0.05 (0.03) *   
#>       16          2       108 (0.86)          0.11      0.02  0.03 (0.03)     
#>       17          1       109 (0.87)          0.06      0.02  0.02 (0.03)     
#>       18         16       125 (1.00)          1.00      0.02               -  
#> -----------------------------------------------------------------------------
#>  Left censoring  : 0.09 (11)
#>  Right centoring : 0.00 (0)
#>  # of nodes      : 125
#> 
#>  Moran's I was computed on contemporaneous autocorrelation using 1/geodesic
#>  values. Significane levels  *** <= .01, ** <= .05, * <= .1.