Indexing diffnet objects (on development)

Access and assign (replace) elements from the adjacency matrices or the vertex attributes data frames.

# S3 method for diffnet
[[(x, name, as.df = FALSE)

# S3 method for diffnet
[[(x, i, j) <- value

# S3 method for diffnet
[(x, i, j, k, drop = FALSE)

# S3 method for diffnet
[(x, i, j, k) <- value

Arguments

x
A diffnet class object.

name
String vector. Names of the vertices attributes.

as.df
Logical scalar. When TRUE returns a data frame, otherwise a list of length \(T\).

i
Index of the i-th row of the adjacency matrix (see details).

j
Index of the j-th column of the adjacency matrix (see details)

value
Value to assign (see details)

k
Index of the k-th slice of the adjacency matrix (see details).

drop
Logical scalar. When TRUE returns an adjacency matrix, otherwise a filtered diffnet object.

Value

In the case of the assigning methods, a diffnet object. Otherwise, for [[.diffnet a vector extracted from one of the attributes data frames, and for [.diffnet a list of length length(k) with the corresponding [i,j] elements from the adjacency matrix.

Details

The [[.diffnet methods provides access to the diffnet attributes data frames, static and dynamic. By providing the name of the corresponding attribute, depending on whether it is static or dynamic the function will return either a data frame--static attributes--or a list of these--dynamic attributes. For the assigning method, [[<-.diffnet, the function will infer what kind of attribute is by analyzing the dimensions of value, in particular we have the following possible cases:

Class	Dimension	Inferred

matrix	\(n\times T\)	Dynamic

matrix	\(n\times 1\)	Static

matrix	\((n\times T)\times 1\)	Dynamic

data.frame	\(n\times T\)	Dynamic

data.frame	\(n\times 1\)	Static

data.frame	\((n\times T)\times 1\)	Dynamic

vector	\(n\)	Static

vector	\(n\times T\)	Dynamic

list*	\(T\) data.frames/matrices/vectors	Dynamic

*: With \(n\times 1\) data.frame/matrix or \(n\) length vector.

Other cases will return with error.

In the case of the slices index k, either an integer vector with the positions, a character vector with the labels of the time periods or a logical vector of length T can be used to specify which slices to retrieve. Likewise, indexing vertices works in the same way with the only difference that, instead of time period labels and a logical vector of length T, vertices ids labels and a logical vector of length n should be provided.

When subsetting slices, the function modifies the toa vector as well as the adopt and cumadopt matrices collapsing network tinmming. For example, if a network goes from time 1 to 20 and we set k=3:10, all individuals who adopted prior to time 3 will be set as adopters at time 3, and all individuals who adopted after time 10 will be set as adopters at time 10, changing the adoption and cumulative adoption matrices. Importantly, k have no gaps, and it should be within the graph time period range.

See also

Examples

# Creating a random diffusion network ---------------------------------------
set.seed(111)
graph <- rdiffnet(100,5)

# Accessing to a static attribute
graph[["real_threshold"]]

#>   [1] 0.54853665 0.84659582 0.03114472 0.07245070 0.41045318 0.04899121
#>   [7] 0.84879455 0.48346461 0.55945227 0.93160364 0.08700101 0.16172018
#>  [13] 0.64275970 0.53030147 0.61148225 0.43890574 0.34671120 0.47003960
#>  [19] 0.23485212 0.51745376 0.33616082 0.40818557 0.56359708 0.35656975
#>  [25] 0.94412400 0.57896703 0.93076293 0.70306857 0.50769567 0.44329401
#>  [31] 0.22484389 0.46930554 0.24102896 0.92141657 0.36478836 0.53217597
#>  [37] 0.79009153 0.03022694 0.41604330 0.50871287 0.62772265 0.62940888
#>  [43] 0.76003392 0.59570532 0.21435065 0.02779840 0.69219762 0.11125828
#>  [49] 0.27786081 0.23251052 0.53142227 0.13572157 0.14677117 0.62526613
#>  [55] 0.62746336 0.38213776 0.01765124 0.67658188 0.16976287 0.73144647
#>  [61] 0.29783417 0.86156361 0.79600628 0.89851621 0.28597520 0.66183314
#>  [67] 0.68525874 0.85841346 0.43467214 0.43923069 0.71706635 0.77696000
#>  [73] 0.16200947 0.58437709 0.66896927 0.62710499 0.33566139 0.09885920
#>  [79] 0.24010096 0.43081253 0.79117821 0.77854059 0.54898132 0.03158283
#>  [85] 0.23706937 0.67690766 0.35349381 0.50029774 0.59047181 0.87296546
#>  [91] 0.91826455 0.53866236 0.57742728 0.17532995 0.99828338 0.79356386
#>  [97] 0.01414975 0.37677331 0.74589013 0.47860022

# Accessing to subsets of the adjacency matrix
graph[1,,1:3, drop=TRUE]

#> $`1`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                               
#> [1,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                         
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . .
#> 
#> $`2`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                               
#> [1,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                         
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . .
#> 
#> $`3`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                               
#> [1,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                         
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . .
#> 

graph[,,1:3, drop=TRUE][[1]]

#> 100 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                               
#>   [1,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [2,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [3,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [4,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [5,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [6,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [7,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [8,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [9,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [10,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [11,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [12,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [13,] . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [14,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [15,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [16,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [17,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [18,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [19,] . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . .
#>  [20,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [21,] . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [22,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [23,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [24,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [25,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [26,] . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . .
#>  [27,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [28,] . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [29,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [30,] . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . .
#>  [31,] . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [32,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [33,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [34,] . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [35,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [36,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [37,] . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [38,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [39,] . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . .
#>  [40,] . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . .
#>  [41,] . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . .
#>  [42,] . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . .
#>  [43,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [44,] . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . .
#>  [45,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [46,] . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . .
#>  [47,] . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [48,] . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [49,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [50,] . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . .
#>  [51,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . .
#>  [52,] . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . .
#>  [53,] . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . .
#>  [54,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [55,] . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [56,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [57,] . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . .
#>  [58,] . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . .
#>  [59,] . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . .
#>  [60,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [61,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [62,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [63,] . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . .
#>  [64,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [65,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [66,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [67,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [68,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [69,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [70,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [71,] . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [72,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [73,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . .
#>  [74,] . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . .
#>  [75,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [76,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . .
#>  [77,] . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . .
#>  [78,] . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [79,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [80,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [81,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [82,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [83,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [84,] . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [85,] . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . .
#>  [86,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [87,] . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [88,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [89,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [90,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [91,] . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . .
#>  [92,] . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . .
#>  [93,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [94,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [95,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . .
#>  [96,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [97,] . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . .
#>  [98,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [99,] . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . .
#> [100,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#>   [1,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [2,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [3,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [4,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [5,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [6,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [7,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [8,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [9,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [10,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [11,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [12,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [13,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [14,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [15,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . .
#>  [16,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [17,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [18,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [19,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [20,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [21,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [22,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [23,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [24,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [25,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [26,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [27,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [28,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [29,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [30,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [31,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [32,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [33,] . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [34,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [35,] . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . .
#>  [36,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [37,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [38,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [39,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [40,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [41,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [42,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [43,] . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . .
#>  [44,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [45,] . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [46,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [47,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [48,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [49,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [50,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [51,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [52,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [53,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [54,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [55,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [56,] . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#>  [57,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [58,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [59,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [60,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [61,] . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [62,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [63,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [64,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [65,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [66,] . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#>  [67,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [68,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [69,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [70,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [71,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [72,] . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#>  [73,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [74,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [75,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [76,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [77,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [78,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [79,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [80,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [81,] . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . .
#>  [82,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [83,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . .
#>  [84,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [85,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [86,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [87,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [88,] . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [89,] . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . .
#>  [90,] . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . .
#>  [91,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [92,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [93,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#>  [94,] . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#>  [95,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [96,] . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . .
#>  [97,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [98,] . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . .
#>  [99,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> [100,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                               
#>   [1,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [2,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [3,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [4,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [5,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [6,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [7,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [8,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>   [9,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [10,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [11,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [12,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [13,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [14,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [15,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [16,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [17,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [18,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [19,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [20,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [21,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [22,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [23,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [24,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [25,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [26,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [27,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [28,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [29,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [30,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [31,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [32,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [33,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [34,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [35,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [36,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [37,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [38,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [39,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [40,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [41,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [42,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [43,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [44,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [45,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [46,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [47,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [48,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [49,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [50,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [51,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [52,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [53,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [54,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [55,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [56,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [57,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [58,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [59,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [60,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [61,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [62,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [63,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [64,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [65,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [66,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [67,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [68,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [69,] . . . . . . . . . . 1 . . . . . . . . . . . . . . . . .
#>  [70,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [71,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [72,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [73,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [74,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [75,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [76,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [77,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [78,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [79,] . . . . 1 . . . . . . . . . . . . . . . . . . . . . . .
#>  [80,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [81,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [82,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [83,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [84,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [85,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [86,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [87,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [88,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [89,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [90,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [91,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [92,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [93,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [94,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [95,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [96,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [97,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [98,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>  [99,] . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> [100,] . . . . . . . . . . . . . . . . . . . . . . . . 1 . . .

# ... Now, as diffnet objects (the default)
graph[1,,1:3, drop=FALSE]

#> Dynamic network of class -diffnet-
#>  Name               : A diffusion network
#>  Behavior           : Random contagion
#>  # of nodes         : 1 (1)
#>  # of time periods  : 3 (1 - 3)
#>  Type               : directed
#>  Final prevalence   : 0.00
#>  Static attributes  : real_threshold (1)
#>  Dynamic attributes : -

graph[,,1:3, drop=FALSE]

#> Dynamic network of class -diffnet-
#>  Name               : A diffusion network
#>  Behavior           : Random contagion
#>  # of nodes         : 100 (1, 2, 3, 4, 5, 6, 7, 8, ...)
#>  # of time periods  : 3 (1 - 3)
#>  Type               : directed
#>  Final prevalence   : 0.09
#>  Static attributes  : real_threshold (1)
#>  Dynamic attributes : -

# Changing values in the adjacency matrix
graph[1, , , drop=TRUE]

#> $`1`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                               
#> [1,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                         
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . .
#> 
#> $`2`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                               
#> [1,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                         
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . .
#> 
#> $`3`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                               
#> [1,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                         
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . .
#> 
#> $`4`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                               
#> [1,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                         
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . .
#> 
#> $`5`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                               
#> [1,] 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                         
#> [1,] . . . . . . . . . . . . . . . . . . . . . . . . . .
#> 

graph[1,,] <- -5
graph[1, , , drop=TRUE]

#> $`1`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#> 
#> $`2`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#> 
#> $`3`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#> 
#> $`4`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#> 
#> $`5`
#> 1 x 100 sparse Matrix of class "dgCMatrix"
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#>                                                                                
#> [1,] -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5 -5
#> 

# Adding attributes (dynamic) -----------------------------------------------
# Preparing the data
set.seed(1122)
x <- rdiffnet(30, 5, seed.p.adopt=.15)

# Calculating exposure, and storing it diffe
expoM <- exposure(x)
expoL <- lapply(seq_len(x$meta$nper), function(x) expoM[,x,drop=FALSE])
expoD <- do.call(rbind, expoL)

# Adding data (all these are equivalent)
x[["expoM"]] <- expoM
x[["expoL"]] <- expoL
x[["expoD"]] <- expoD

# Lets compare
identical(x[["expoM"]], x[["expoL"]]) # TRUE

#> [1] TRUE

identical(x[["expoM"]], x[["expoD"]]) # TRUE

#> [1] TRUE