Difference in Time of Adoption (TOA) between individuals

Creates \(n \times n\) matrix indicating the difference in times of adoption between each pair of nodes

toa_diff(obj, t0 = NULL, labels = NULL)

Arguments

obj
Either an integer vector of size \(n\) containing time of adoption of the innovation, or a diffnet object.

t0
Integer scalar. Sets the lower bound of the time window (e.g. 1955).

labels
Character vector of size \(n\). Labels (ids) of the vertices.

Value

An \(n \times n\) symmetric matrix indicating the difference in times of adoption between each pair of nodes.

Details

Each cell ij of the resulting matrix is calculated as \(toa_j - toa_i\), so that whenever its positive it means that the j-th individual (alter) adopted the innovation sooner.

Examples

# Generating a random vector of time
set.seed(123)
times <- sample(2000:2005, 10, TRUE)

# Computing the TOA differences
toa_diff(times)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    3    0   -1   -1    3    0    2    1     3
#>  [2,]   -3    0   -3   -4   -4    0   -3   -1   -2     0
#>  [3,]    0    3    0   -1   -1    3    0    2    1     3
#>  [4,]    1    4    1    0    0    4    1    3    2     4
#>  [5,]    1    4    1    0    0    4    1    3    2     4
#>  [6,]   -3    0   -3   -4   -4    0   -3   -1   -2     0
#>  [7,]    0    3    0   -1   -1    3    0    2    1     3
#>  [8,]   -2    1   -2   -3   -3    1   -2    0   -1     1
#>  [9,]   -1    2   -1   -2   -2    2   -1    1    0     2
#> [10,]   -3    0   -3   -4   -4    0   -3   -1   -2     0